+++ legadilo/feeds/models/__init__.py
@@ -17,6 +17,7 @@
 from .feed import Feed
 from .feed_article import FeedArticle
 from .feed_category import FeedCategory
+from .feed_deleted_article import FeedDeletedArticle
 from .feed_tag import FeedTag
 from .feed_update import FeedUpdate
 
@@ -24,6 +25,7 @@ __all__ = [
     "Feed",
     "FeedArticle",
     "FeedCategory",
+    "FeedDeletedArticle",
     "FeedTag",
     "FeedUpdate",
 ]